+++ ivyIpcamSdk/src/commonMain/kotlin/com/ivyiot/ipcam_sdk/utils/BytesPerSecond.kt
@@ -3,9 +3,9 @@ package com.ivyiot.ipcam_sdk.utils
 import kotlin.math.round
 import kotlin.math.pow
 
-data class BytesPerSecond(val value: Int) {
+data class BytesPerSecond(val value: UInt) {
     fun getFormattedValue(): String {
-        var speed = value.toFloat()
+        var speed = value.toDouble()
         var unit = "B"
 
         if (speed >= 1024) {
@@ -23,7 +23,7 @@ data class BytesPerSecond(val value: Int) {
             unit = "GB"
         }
 
-        val decimalCount = if (speed >= 100.0f || unit == "B") 0 else 1
+        val decimalCount = if (speed >= 100.0 || unit == "B") 0 else 1
         val decimalShiftValue =  10.0.pow(decimalCount)
         return "${round(speed * decimalShiftValue) / decimalShiftValue} ${unit}/S"
     }
